Check that all Lua files are present before doing build tests
