+++ configur.lua
@@ -1,4 +1,22 @@
-require("lua.args")
+require("lua.boot")
+
+-- Determine operating system family
+System = CheckSystemFamily()
+
+-- Check 'lua' contains required files
+CheckDirContains("lua", {
+	"asm.lua",
+	"checks.lua",
+	"compiler.lua",
+	"compiler/custom.lua",
+	"compiler/gcc.lua",
+	"compiler/wcc.lua",
+	"makefile.lua",
+	"util.lua",
+	}
+)
+
+-- Required files are found and can now be loaded
 require("lua.checks")
 require("lua.makefile")
 require("lua.util")
@@ -7,19 +25,27 @@ Target = {}
 GetOpt()
 
 -- Check parameters and which makefile to generate
-
 Target.makefile = CheckMakefileRequestValid()
 
--- Can the system achieve basic things?
-System = CheckSystemFamily()
+-- Create a unique name for testing file and folder
 TmpFolder = UniqueName()
 
+-- Check creating/deleting files and folders works
 CheckCreateDirCmd(TmpFolder)
 CheckRemoveFileCmd(TmpFolder)
 CheckRemoveDirCmd(TmpFolder)
 
--- Does the project directory look correct?
-CheckDirContains("inc", {"net/if.h", "tcp.h"}) -- TODO: Add all the important headers in a table and call in place of this inline variable
+-- Check 'inc' contains required files (for application to include)
+CheckDirContains("inc", {
+	"arpa/inet.h",
+	"netdb.h",
+	"netinet/in.h",
+	"sys/socket.h",
+	"tcp.h",
+	}
+)
+
+-- Check 'src' contains required files
 CheckDirContains("src", MakefileCoreSource())
 
 -- Create a basic C file to test the compiler
