+++ lua/compiler/wcc.lua
@@ -1,3 +1,8 @@
+--[[
+	wcc.lua contains the functions needed to test
+	Open Watcoms 1.9 or later C compilers (wcc and wcc386)
+]]
+
 local function CheckWccCompiler(tmpName)
 	Compiler.type = "watcom"
 	Compiler.output = "-fo="
